requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==2.3.7
numpy==1.24.3
python-dotenv==1.0.0
email-validator==2.1.0.post1
SQLAlchemy==2.0.21
alembic==1.12.0
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.5
MarkupSafe==2.1.3